pyproject.toml
[project]
name = "vsc-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "allure-pytest>=2.15.3",
    "playwright>=1.58.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
addopts = "--alluredir=report --junitxml=report.xml"
markers = [
    "test_id: Test Case ID for reporting"
]
log_cli = true
log_file = "test.log"
log_file_level = "INFO"
